Update dist folder
