+++ src/modules/showNotification.js
@@ -1,0 +1,17 @@
+const showNotification = (message) => {
+  const notification = document.createElement('div');
+  notification.classList.add('notification');
+  notification.textContent = message;
+  document.body.appendChild(notification);
+  setTimeout(() => {
+    notification.classList.add('show');
+  }, 50);
+  setTimeout(() => {
+    notification.classList.remove('show');
+    setTimeout(() => {
+      notification.remove();
+    }, 300);
+  }, 2000);
+};
+
+export default showNotification;
